+++ src/test/java/lotto/model/RankTest.java
@@ -1,10 +1,9 @@
-package lotto.util;
+package lotto.model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class RankTest {
     @DisplayName("로또 번호를 맞힌 개수대로 등수를 반환한다.")
